Combine track slot executors
