+++ src/celeritas/optical/action/detail/PropagateExecutor.hh
@@ -32,11 +32,6 @@ struct PropagateExecutor
 //---------------------------------------------------------------------------//
 CELER_FUNCTION void PropagateExecutor::operator()(CoreTrackView& track)
 {
-    if (track.surface_physics().is_crossing_boundary())
-    {
-        return;
-    }
-
     auto&& sim = track.sim();
     CELER_ASSERT(sim.status() == TrackStatus::alive);
